requirements.txt
langchain
langchain-text-splitters
fastapi==0.109.0
uvicorn==0.27.0
litellm==1.17.9
pydantic==2.5.3
setuptools
requests
python-dotenv
PyJWT
google-generativeai